# smsc/xpmem rcache: refactor to use tree find() instead of iterate()