@gkatev 2023-01-18
Hi @devreal, I plan to work again on it soon. The `find() instead of iterate()` part is functional, but I want to verify one last time that it all works correctly (thus the debug prints). And I still want to look into making endpoints be 1 per process. I also had thread safety concerns but these might be out of the scope of this PR.

Any feedback is welcome. For example:
> I took "bound" to mean the last valid byte in the range, inclusive

Would you say this is accurate/desired?

Or if you meant code contributions, one could look into the -/+1 stuff in rcache/interval_tree code, or maybe into the 1 ep/proc part (creating a proc-to-ep structure, so as to have all components use the same endpoint/apid/regcache).