@devreal 2023-01-18
I suggest to split this PR into multiple pieces that are related (the `find` part, the flags, and endpoints). That makes reviewing easier and we can discuss the different points in separate issues. Right now it's a bit fuzzy. Splitting would also avoid the situation where fixes that are ready are held up by work in progress.