@gkatev 2023-01-30
I created #11358 to replace this PR. It includes the implementation of find() instead of iterate(). It also contains the implementation of the extra rcache flags in a separate commit -- if we'd like a different PR for that too, they are easy to separate.

Besides the changes in #11358, there pends the enhancement of having one rcache per process (instead of per endpoint), which is currently unimplemented and not contained in any PR.

There also some atomicity concerns, but they are likely out of the context of these changes, and I'm not currently looking into them.